requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cytoolz==0.11.2
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
Flask==2.0.3
frozenlist==1.3.0
gunicorn==20.1.0
hexbytes==0.2.2
idna==3.3
install==1.3.5
ipfshttpclient==0.8.0a2
itsdangerous==2.1.1
Jinja2==3.0.3
jsonschema==3.2.0
lru-dict==1.1.7
MarkupSafe==2.1.0
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.22.3
pandas==1.4.1
parsimonious==0.8.1
protobuf==3.19.4
pycryptodome==3.14.1
pymongo==4.0.2
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rlp==2.0.1
six==1.16.0
toolz==0.11.2
urllib3==1.26.8
varint==1.0.2
web3==5.28.0
websockets==9.1
Werkzeug==2.0.3
yarl==1.7.2
